+++ ProyectosNetbeansWS/TouresBalonProductosWS/src/main/java/aes/pica/touresbalon/touresbalonproductosws/entidades/productos/TarifaTransporte.java
@@ -1,5 +1,5 @@
 package aes.pica.touresbalon.touresbalonproductosws.entidades.productos;
-// Generated 3/10/2016 11:01:42 PM by Hibernate Tools 4.3.1
+// Generated Oct 16, 2016 3:40:40 PM by Hibernate Tools 4.3.1
 
 
 import java.math.BigDecimal;
@@ -14,7 +14,7 @@ public class TarifaTransporte  implements java.io.Serializable {
 
      private int idTransporte;
      private String nombreTransporte;
-     private double precio;
+     private BigDecimal precio;
      private Set productos = new HashSet(0);
 
     public TarifaTransporte() {
@@ -24,7 +24,7 @@ public class TarifaTransporte  implements java.io.Serializable {
     public TarifaTransporte(int idTransporte) {
         this.idTransporte = idTransporte;
     }
-    public TarifaTransporte(int idTransporte, String nombreTransporte, double precio, Set productos) {
+    public TarifaTransporte(int idTransporte, String nombreTransporte, BigDecimal precio, Set productos) {
        this.idTransporte = idTransporte;
        this.nombreTransporte = nombreTransporte;
        this.precio = precio;
@@ -45,11 +45,11 @@ public class TarifaTransporte  implements java.io.Serializable {
     public void setNombreTransporte(String nombreTransporte) {
         this.nombreTransporte = nombreTransporte;
     }
-    public double getPrecio() {
+    public BigDecimal getPrecio() {
         return this.precio;
     }
     
-    public void setPrecio(double precio) {
+    public void setPrecio(BigDecimal precio) {
         this.precio = precio;
     }
     public Set getProductos() {
